+++ src/Support/Changelog.md
@@ -7,6 +7,7 @@
 
 - Dropped PHP 8.0 & PHP 8.1 Support
 - Dropped Laravel 9 Support
+- Added the `Features` accessor class, which is a syntactic sugar for areas of the configuration
 
 ## 3.x Series