+++ src/Support/Changelog.md
@@ -9,7 +9,7 @@
 - Dropped Laravel 9 Support
 - Added PHP 8.3 Support
 - Added Laravel 11 Support
-- Changed minimum Laravel version to v10.38.2
+- Changed minimum Laravel version to v10.43
 - Added the `Features` accessor class, which is a syntactic sugar for areas of the configuration
 - Added the `Merchant` DTO class, that complies with the common, v4 `Merchant` interface
 - Added the `Address` DTO class (implements the `Address` interface)