+++ src/Support/Changelog.md
@@ -5,6 +5,7 @@
 ## Unreleased
 ##### 2024-XX-YY
 
+- Changed the minimal Address module requirement to v3.4.1
 - Changed the `ConfigurableModel::hasConfiguration()` trait to return false on empty arrays as well (until now, only `null` was considered as no config)
 - Added the search engine and inventory feature configuration wrappers